# docs/swagger.yaml
definitions:
...
  model.Transaction:
    properties:
      amount:
        type: number
      categoryID:
        description: Foreign key to Category
        type: string
      created_at:
        type: string
      date:
        type: string
      deleted_at:
        description: Soft delete
        type: string
      description:
        type: string
      id:
        description: Adds some metadata fields to the table
        type: string
      updated_at:
        type: string
      userID:
        description: Foreign key to User
        type: string
    type: object
  model.TransactionGroupedByCategory:
    properties:
      category_name:
        type: string
      category_type:
        type: string
      transactions:
        items:
          $ref: '#/definitions/model.Transaction'
        type: array
    type: object
  model.User:
    properties:
      created_at:
        type: string
      deleted_at:
        description: Soft delete
        type: string
      email:
        description: Email field with JSON and unique constraint
        type: string
      first_name:
        type: string
      id:
        description: Adds some metadata fields to the table
        type: string
      last_name:
        type: string
      password:
        description: Only for email/password login
        type: string
      refreshToken:
        description: To store the refresh token
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/statistics/category:
    get:
      description: Returns income and expense statistics by category and date range
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category ID
        in: query
        name: category_id
        required: true
        type: string
      - description: Start Date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End Date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Transaction'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get statistics by category
      tags:
      - statistics
  /api/transaction:
    get:
      consumes:
      - application/json
      description: Retrieve transactions by category and date range
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category ID
        in: query
        name: categoryId
        type: string
      - description: Start Date in YYYY-MM-DD format
        in: query
        name: startDate
        type: string
      - description: End Date in YYYY-MM-DD format
        in: query
        name: endDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.TransactionGroupedByCategory'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get transactions grouped by category
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Adds an income or expense transaction by category
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Transaction Data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/model.Transaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Transaction'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Add a new transaction
      tags:
      - transactions
  /api/user:
    get:
      consumes:
      - application/json
      description: Get all existing users
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Create a User
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Name
        in: body
        name: name
        required: true
        schema:
          type: string
      - description: Email
        in: body
        name: email
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
      tags:
      - Users
  /api/user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Get one user by ID
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
      tags:
      - user
...
swagger: "2.0"